auth_service/requirements.txt
Django==4.0.3
djangorestframework==3.13.1
PyJWT==2.3.0
django-cors-headers==3.11.0
psycopg2==2.9.3